Better detection of failed forked subprocess in #run_ruby_script! & #run!
In the particular reproducible failure, the system had a small amount of
RAM with no swap enabled leading to various `Errno::ENOMEM` and signal-9
termination of parent and child processes.

The implementation of Thor's `#run` and `#run_ruby_script` use
`Kernel.system` when `:capture => false`. It turns out that
`Kernel.system` returns `nil` if the forked subprocess had an issue
(like no RAM available to fork/exec) which did not update the `$?` Ruby
global variable, so `$?` remained `nil`.

As of this commit, there is detection of a `nil` `$?` value which will
warn the user and crash a bit louder (using the fatal `"!!!!!!!"`
message format).

Additionally, if the subprocess raises an exception (which is the
behavior of `Kernel.\`` when `:capture => true`), the raised exception
will be caught, a fatal message issued, and exception is re-raised to
crash Busser.

The goal for this commit is to be clearer with the end user that Busser
is crashing for a good reason--in the original filed issue, it was
memory constraints.

Closes #17
References test-kitchen/busser-serverspec#22