Add http_proxy support to `busser plugin install`.
This change will detect and use either the `http_proxy` or the
`HTTP_PROXY` environment variable when performing the plugin
installations, which delegate into RubyGems `gem install` code. This
delivers the functional equivalent behavior of running:

    gem install busser-<PLUGIN> --http-proxy http://proxy.local:8123

As the RubyGems install code had to be revisited, an alternative
implementation that is much simpler and (anecdotally) faster was used.
Using `Gem::Dependency.new(name, version).matching_specs` means we can
avoid any network calls to see if any installed plugin version satisfies
the version requirement. There is a small change in behavior where prior
to this the latest version was always checked, but given Busser's role
as a helper on an ephemeral testing node, there shouldn't be a pressing
need to update a plugin version (i.e. "if you ask for it to be installed
and there is any version, use it!").

Finally, a new optional flag `--verbose` is added to the
`busser plugin install` command which helps output the http_proxy server
used (if any) and the HTTP calls to the RubyGems API services. A future
version of Test Kitchen may use this flag when it is invoked with a
debug log level.